integers required

+++ applications/RomApplication/python_scripts/hrom_training_utility.py
@@ -83,7 +83,7 @@ class HRomTrainingUtility(object):
                 candidate_ids = np.r_[candidate_ids, np.array(this_modelpart_condition_ids)+number_of_elements]
 
         if np.size(candidate_ids)>0:
-            self.candidate_ids = np.unique(candidate_ids) - 1 # this -1 takes into account the id difference in numpy and Kratos
+            self.candidate_ids = np.unique(candidate_ids).astype(int) - 1 # this -1 takes into account the id difference in numpy and Kratos
         else:
             self.candidate_ids = None
